+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/AttestationGossipManager.java
@@ -13,21 +13,15 @@
 
 package tech.pegasys.teku.networking.eth2.gossip;
 
-import com.google.common.base.Throwables;
-import io.libp2p.pubsub.MessageAlreadySeenException;
-import io.libp2p.pubsub.NoPeersForOutboundMessageException;
-import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import org.hyperledger.besu.plugin.services.metrics.Counter;
 import org.hyperledger.besu.plugin.services.metrics.LabelledMetric;
 import tech.pegasys.teku.infrastructure.metrics.TekuMetricCategory;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.networking.eth2.gossip.subnets.AttestationSubnetSubscriptions;
 import tech.pegasys.teku.spec.datastructures.attestation.ValidateableAttestation;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
-import tech.pegasys.teku.spec.datastructures.operations.AttestationData;
 
 public class AttestationGossipManager implements GossipManager {
   private static final Logger LOG = LogManager.getLogger();
@@ -37,8 +31,7 @@ public class AttestationGossipManager implements GossipManager {
   private final Counter attestationPublishSuccessCounter;
   private final Counter attestationPublishFailureCounter;
 
-  private UInt64 lastErroredSlot;
-  private Throwable lastRootCause;
+  private final GossipFailureLogger gossipFailureLogger = new GossipFailureLogger("attestation");
 
   public AttestationGossipManager(
       final MetricsSystem metricsSystem,
@@ -69,38 +62,9 @@ public class AttestationGossipManager implements GossipManager {
               attestationPublishSuccessCounter.inc();
             },
             error -> {
-              logWithSuppression(error, attestation);
+              gossipFailureLogger.logWithSuppression(error, attestation.getData().getSlot());
               attestationPublishFailureCounter.inc();
             });
-  }
-
-  synchronized void logWithSuppression(final Throwable error, final Attestation attestation) {
-    final AttestationData attestationData = attestation.getData();
-    final Throwable rootCause = Throwables.getRootCause(error);
-
-    final boolean suppress =
-        attestationData.getSlot().equals(lastErroredSlot)
-            && rootCause.getClass().equals(lastRootCause.getClass());
-
-    lastErroredSlot = attestationData.getSlot();
-    lastRootCause = rootCause;
-
-    if (lastRootCause instanceof MessageAlreadySeenException) {
-      LOG.debug(
-          "Failed to publish attestation(s) for slot {} because the message has already been seen",
-          lastErroredSlot);
-    } else if (lastRootCause instanceof NoPeersForOutboundMessageException) {
-      LOG.log(
-          suppress ? Level.DEBUG : Level.WARN,
-          "Failed to publish attestations(s) for slot {} because no peers were available on the required gossip topic",
-          lastErroredSlot);
-    } else {
-      LOG.log(
-          suppress ? Level.DEBUG : Level.ERROR,
-          "Failed to publish attestation(s) for slot {}",
-          lastErroredSlot,
-          error);
-    }
   }
 
   public void subscribeToSubnetId(final int subnetId) {